requirements.txt
Django==2.2.16
djangorestframework==3.12.2
pytz==2022.7
python-decouple==3.3
djangorestframework-simplejwt
PyJWT==1.7.1
drf-yasg==1.20.0
django-solo==1.1.5
django-unique-upload==0.2.1
psycopg2-binary==2.9.5
PyJWT==1.7.1
django-filter==2.3.0
django-import-export==2.4.0
django-cors-headers==3.1.0
django-modeltranslation==0.16.1
mysqlclient==1.4.4
Pillow==8.1.0